Move iter_items to RocksDBStorePartition
