+++ quixstreams/state/rocksdb/windowed/partition.py
@@ -1,7 +1,5 @@
 import logging
-from typing import Iterator, Optional, Tuple, cast
-
-from rocksdict import ReadOptions
+from typing import Iterator, Optional, cast
 
 from quixstreams.state.exceptions import ColumnFamilyDoesNotExist
 from quixstreams.state.recovery import ChangelogProducer
@@ -48,17 +46,6 @@ class WindowedRocksDBStorePartition(RocksDBStorePartition):
         self._ensure_column_family(LATEST_TIMESTAMPS_CF_NAME)
         self._ensure_column_family(GLOBAL_COUNTER_CF_NAME)
         self._ensure_column_family(VALUES_CF_NAME)
-
-    def iter_items(
-        self, from_key: bytes, read_opt: ReadOptions, cf_name: str = "default"
-    ) -> Iterator[Tuple[bytes, bytes]]:
-        cf = self.get_column_family(cf_name=cf_name)
-
-        # RDict accept Any type as value but we only write bytes so we should only get bytes back.
-        return cast(
-            Iterator[Tuple[bytes, bytes]],
-            cf.items(from_key=from_key, read_opt=read_opt),
-        )
 
     def iter_keys(self, cf_name: str = "default") -> Iterator[bytes]:
         """
